Update main.js
Created 3 classes for each current component type. These are:
- Component
- ColorComp
- ImageComp

Image comp is for drawing rectangles with an image as background, ColorComp is for drawing a colored rectangle, and Component
is the parent class and should not be used in most cases. In addition, collision and interaction booleans were added to the component properties. Collision only acts on an object if its collision property is true. hasInteract doesn't do anything yet.