Update main.js
Changed left and right movement using A and D to a static speed setter of 10 or -10.
Added gravity, natural field on component creation. If a component is assigned a newPos() gravity will be applied.
Added a jump using W key. Pressing it while above an object (this.relative === "above"), or on bottom of map will set speedY to -25, pushing you upwards

Jump may or may not stay in the game, final movement will be determined based on later mechanics